+++ components/SelectedNodeComponent.tsx
@@ -1,8 +1,21 @@
 import React, { useEffect, useState } from 'react';
 import { Handle, NodeProps, Position } from 'reactflow';
 import { SelectedNodeData } from '../model';
-
+import { Session } from '../store/session';
+import { useStores } from '../store/store-container';
 type PineNodeProps = NodeProps<SelectedNodeData>;
+
+const onSelectedNodeClick = (session: Session, alias: string) => {
+  session.setContext(alias);
+};
+
+const onCandidateColumnClick = (session: Session, column: string, type: 'select' | 'order') => {
+  if (type === 'select') {
+    session.appendAndUpdateExpression(`${column}, `);
+  } else {
+    session.appendAndUpdateExpression(`${column} desc, `);
+  }
+};
 
 const TableNode = ({
   order,
@@ -10,13 +23,17 @@ const TableNode = ({
   schema,
   color,
   alias,
+  sessionId,
 }: {
   order: number;
   table: string;
   schema: string;
   color?: string | null;
   alias: string;
+  sessionId: string;
 }) => {
+  const { global } = useStores();
+  const session = global.getSession(sessionId);
   return (
     <div
       style={{
@@ -28,12 +45,14 @@ const TableNode = ({
     >
       {/* Node */}
       <div
+        onClick={() => onSelectedNodeClick(session, alias)}
         style={{
           position: 'relative',
           padding: '12px 10px 5px 10px',
           border: '2px solid black',
           background: 'lightgray',
           borderRadius: '5px',
+          cursor: 'pointer',
         }}
       >
         {/* Table */}
@@ -153,51 +172,67 @@ const SelectedColumns = ({ columns }: { columns: string[] }) => (
   </div>
 );
 
-const CandidateColumns = ({ columns }: { columns: string[] }) => (
-  <div
-    style={{
-      maxHeight: columns.length > 0 ? '500px' : '0',
-      overflow: 'hidden',
-      width: '100%',
-    }}
-  >
+const CandidateColumns = ({
+  columns,
+  sessionId,
+  type,
+}: {
+  columns: string[];
+  sessionId: string;
+  type: 'select' | 'order';
+}) => {
+  const { global } = useStores();
+  const session = global.getSession(sessionId);
+  return (
     <div
       style={{
-        display: 'flex',
-        flexWrap: 'wrap',
-        gap: '4px',
-        justifyContent: 'center',
+        maxHeight: columns.length > 0 ? '500px' : '0',
+        overflow: 'hidden',
         width: '100%',
       }}
     >
-      {columns.map(column => (
-        <div
-          key={column}
-          style={{
-            fontSize: '8px',
-            fontFamily: 'Courier, monospace',
-            background: 'transparent',
-            padding: '2px 6px',
-            borderRadius: '8px',
-            border: '1px solid #ccc',
-            color: '#666',
-          }}
-        >
-          {column}
-        </div>
-      ))}
+      <div
+        style={{
+          display: 'flex',
+          flexWrap: 'wrap',
+          gap: '4px',
+          justifyContent: 'center',
+          width: '100%',
+        }}
+      >
+        {columns.map(column => (
+          <div
+            key={column}
+            onClick={() => onCandidateColumnClick(session, column, type)}
+            style={{
+              fontSize: '8px',
+              fontFamily: 'Courier, monospace',
+              background: 'transparent',
+              padding: '2px 6px',
+              borderRadius: '8px',
+              border: '1px solid #ccc',
+              color: '#666',
+              cursor: 'pointer',
+            }}
+          >
+            {column}
+          </div>
+        ))}
+      </div>
     </div>
-  </div>
-);
+  );
+};
 
 const Columns = ({
   columns,
   suggested,
   type,
+  sessionId,
 }: {
   columns: string[];
   suggested: string[];
   type: 'select' | 'order';
+  sessionId: string;
 }) => {
   const selectedColumnsSet = new Set(columns.filter(Boolean));
   const candidateColumns = suggested.filter(col => !selectedColumnsSet.has(col));
@@ -240,7 +275,7 @@ const Columns = ({
             margin: '8px',
           }}
         >
-          <CandidateColumns columns={candidateColumns} />
+          <CandidateColumns columns={candidateColumns} sessionId={sessionId} type={type} />
         </div>
       )}
     </div>
@@ -258,12 +293,30 @@ const SelectedNodeComponent: React.FC<PineNodeProps> = ({ data }) => {
     orderColumns,
     suggestedColumns,
     suggestedOrderColumns,
+    sessionId,
   } = data;
   return (
     <div>
-      <TableNode order={order} table={table} schema={schema} color={color} alias={alias} />
-      <Columns columns={selectedColumns} suggested={suggestedColumns} type="select" />
-      <Columns columns={orderColumns} suggested={suggestedOrderColumns} type="order" />
+      <TableNode
+        order={order}
+        table={table}
+        schema={schema}
+        color={color}
+        alias={alias}
+        sessionId={sessionId}
+      />
+      <Columns
+        columns={selectedColumns}
+        suggested={suggestedColumns}
+        type="select"
+        sessionId={sessionId}
+      />
+      <Columns
+        columns={orderColumns}
+        suggested={suggestedOrderColumns}
+        type="order"
+        sessionId={sessionId}
+      />
     </div>
   );
 };